[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-int calculation (int x, int y){
-    return (x+y)*(x-y);
+long long calculation (int x, int y){
+    return (long long) (x+y)*(x-y);
 }
